✨ feat(NoInternetConnectionView.swift)
The project file was updated to include the newly added NoInternetConnectionView.swift file. In AdvertisementDetailsViewController, the descriptionLabel and contactsLabel are now set to be visible by default. Additionally, the NoInternetConnectionView is shown when there is a network error in both AdvertisementDetailsViewController and ListOfAdvertisementsViewController. Finally, a new NoInternetConnectionView class was added to display a view indicating that there is no internet connection